+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/BaseControl.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.api.reconciler;
 
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
