+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/Controller.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.api.reconciler;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
