+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/api/reconciler/DeleteControl.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.api;
+package io.javaoperatorsdk.operator.api.reconciler;
 
 public class DeleteControl extends BaseControl<DeleteControl> {
