+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/retry/RetryExecution.java
@@ -2,7 +2,7 @@ package io.javaoperatorsdk.operator.processing.retry;
 
 import java.util.Optional;
 
-import io.javaoperatorsdk.operator.api.RetryInfo;
+import io.javaoperatorsdk.operator.api.reconciler.RetryInfo;
 
 public interface RetryExecution extends RetryInfo {
