+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/api/DeleteControlTest.java
@@ -1,5 +1,6 @@
 package io.javaoperatorsdk.operator.api;
 
+import io.javaoperatorsdk.operator.api.reconciler.DeleteControl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
