+++ operator-framework/src/test/resources/compile-fixtures/AdditionalReconcilerInterface.java
@@ -1,7 +1,7 @@
 package io;
 
 import io.fabric8.kubernetes.client.CustomResource;
-import io.javaoperatorsdk.operator.api.Reconciler;
+import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 import java.io.Serializable;
